+++ app/views/main.hbs
@@ -5,5 +5,8 @@
 </head>
 <body>
   <h1>Welcome, {{user}}</h1>
+  <form method="get" action="/users">
+    <button type="submit">List Users</button>
+  </form>
 </body>
 </html>
